% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legends.R
\name{legend_top}
\alias{legend_top}
\title{Automatic placement legends for plots}
\usage{
legend_top(horiz = TRUE, ...)
}
\arguments{
\item{horiz}{logical; if TRUE, set the legend horizontally rather than
vertically (specifying horiz overrides the ncol specification)}

\item{...}{Further arguments to \code{legend()}}
}
\description{
Automatic placement legends for plots
}
