% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.R
\name{pal}
\alias{pal}
\title{Generate a nice palette of colors}
\usage{
pal(n, alpha = 1)
}
\arguments{
\item{n}{Number of colors to generate, evenly spaced on \code{grDevice::hcl} scale}

\item{alpha}{Partial transparency inherited from \code{grDevice::hcl}}
}
\description{
Generate a character vector of colors by specifying the specifying hue,
chroma and luminance.
}
\examples{
pal(3)

}
