% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_all.R
\name{render_all}
\alias{render_all}
\title{Automatic rendering of Rmarkdown files}
\usage{
render_all(files, knit_root_dir = getwd(), quiet = TRUE, ...)
}
\arguments{
\item{files}{A vector of character values listing the .Rmd, .R, or .md files
to render.}

\item{knit_root_dir}{A character string specifying the knitting directory.
Default is the current working directory with \code{\link[=getwd]{getwd()}}.}

\item{quiet}{An logical option to suprress printing during rendering from
knitr, pandoc command line and others. Default is TRUE.}

\item{...}{Any other \code{\link[rmarkdown:render]{rmarkdown::render()}} parameters.}
}
\description{
Silent rendering of a collection of Rmarkdown files with the current working
directory as the knitting directory.
}
\details{
If \code{files} is not specified, all .Rmd files in the current directory
are located and rendered.
}
\seealso{
\code{\link[rmarkdown:render]{rmarkdown::render()}}
}
